extract shared components
