+++ packages/frontend/src/pages/interop/components/comparison/compareMetric.ts
