+++ packages/frontend/src/pages/interop/components/comparison/types.ts
@@ -1,0 +1,4 @@
+export type ComparisonSide = {
+  label: string
+  color: string
+}
