+++ packages/frontend/src/pages/interop/token-frameworks/components/comparison/types.ts
@@ -1,3 +1,4 @@
+import type { ComparisonSide } from '../../../components/comparison/types'
 import type { InteropTokenFramework } from '../../getInteropTokenFrameworksData'
 
 export type ComparableEntry = {
@@ -10,4 +11,12 @@ export type ComparableEntry = {
 export type Side = {
   framework: InteropTokenFramework
   entry: ComparableEntry
+}
+
+export function toComparisonSide(
+  side: Side | undefined,
+): ComparisonSide | undefined {
+  return side
+    ? { label: side.framework.label, color: side.framework.color }
+    : undefined
 }
